refactor(ai): concern-split seven near-cap planning modules (#4291)
Slice B extracts feedstock/fabric labour helpers, phase dispatch outcomes,
GP peace targets, economy civilian-work and recruit passes, filter caps,
diplomacy bonus scaling, and stalled-peace predicates so all seven
350–393-line planning hosts regain durable headroom under the 400-line gate.

Refs #4291
