+++ packages/colonizethis_ai/lib/src/planning/domain_planner_orchestrator_economy.dart
@@ -1,15 +1,13 @@
 import 'dart:math' as math;
 
-
 import '../perception/perception_snapshot.dart';
 import '../util/orders_builder.dart';
 import 'build_planner.dart';
 import 'domain_planner_orchestrator_economy_build.dart';
+import 'domain_planner_orchestrator_economy_civilian_work.dart';
+import 'domain_planner_orchestrator_economy_recruit.dart';
 import 'economy_phase_gates.dart';
-import 'expand_phase_planner_economy.dart';
 import 'growth_stage.dart';
-import 'growth_stage_builder_relocation.dart';
-import 'growth_stage_work_priorities.dart';
 import 'goal_manager.dart';
 import 'phase_planner_dispatch.dart';
 import 'phase_planner_economy_filter.dart';
@@ -20,6 +18,9 @@ import 'planning_helpers.dart'
         isAtWarWithAnyGreatPower,
         isPursuingTechStealPosture;
 import 'planning_imports.dart';
+
+export 'domain_planner_orchestrator_economy_recruit.dart'
+    show appendEconomyPeasantRecruit;
 
 final _log = packageLogger('domain_planner_orchestrator_economy');
 
@@ -62,11 +63,6 @@ EconomyDomainPlannersResult runEconomyDomainPlanners({
     snapshot: snapshot,
   );
   final growthStagePlannerEnabled = ctx.growthStagePlannerEnabled;
-  // Refs #3288: accumulate the orchestrator-emitted economy families (work,
-  // recruit, build) into a single mutable [OrdersBuilder] and freeze once,
-  // replacing the prior chained `Orders.copyWith` appends that allocated an
-  // intermediate immutable [Orders] per family. Read-only suggestion calls
-  // observe the accumulated state via the builder's cached snapshot.
   final ordersBuilder = OrdersBuilder.from(ctx.orders);
   final domainWeights = ctx.domainWeights;
 
@@ -86,21 +82,8 @@ EconomyDomainPlannersResult runEconomyDomainPlanners({
     ctx.game,
     ctx.topology,
     ordersBuilder.build(),
-    // Refs #3793 live wiring: enumerate civilian build candidates only when the
-    // civilian build planner is enabled. Default-off keeps the candidate pool
-    // byte-identical to the military+naval path (SPEC ACWire1).
     includeCivilianBuilds: ctx.civilianBuildPlannerEnabled,
   );
-  // Refs #3793 live wiring: build the civilian scoring input once per turn so
-  // the build pass can apply the min-cap floor, replacement urgency, phase
-  // multiplier, and Spy demand boost. `null` when the planner is disabled,
-  // leaving `pickBuildOrder` inert (SPEC § Live economy wiring).
-  // Refs #3793 decision #10: the Spy demand boost fires when the GP is at war
-  // OR pursuing a tech-steal posture (behind the most-advanced rival GP's
-  // unlocked-tech count by `kCivilianBuildSpyTechStealDeficit`). This replaces
-  // the prior war-posture-only approximation so passive spy RP value is reflected
-  // even at peace (SPEC § Live economy wiring; AC4c). Inert unless
-  // `civilianBuildPlannerEnabled` (buildCivilianBuildScoringInput returns null).
   final spyDemand =
       isAtWarWithAnyGreatPower(ctx.game, snapshot) ||
       isPursuingTechStealPosture(ctx.game, ctx.nationId);
@@ -108,10 +91,6 @@ EconomyDomainPlannersResult runEconomyDomainPlanners({
     ctx: ctx,
     phaseName: phasePlan.phase.name,
     spyDemand: spyDemand,
-    // Refs #3793 slice 8 (decision #13): the dispatch's continuous
-    // `newWorldCivilian` weight drives smooth phase-multiplier ramping
-    // (hysteresis) so per-phase civilian priorities transition continuously
-    // across the EXPAND→COLONIAL→DEVELOP boundary instead of hard-switching.
     phaseProgress: phasePlan.priorityWeights.newWorldCivilian,
   );
   final hasSpyWork = workCandidates.any(
@@ -120,42 +99,8 @@ EconomyDomainPlannersResult runEconomyDomainPlanners({
   var workThreshold =
       40 -
       (hasSpyWork ? getAgendaSpyOrderModifier(ctx.config.hiddenAgendaId) : 0);
-  // Refs #2509 S5: derive the DEVELOP-phase economy gate from the
-  // dispatched phase plan instead of recomputing
-  // `isObserverDevelopPhase` (which itself recomputes
-  // `observerGoalPhaseFor`) per player turn. The phase dispatcher
-  // already resolved `observerGoalPhaseFor` once via
-  // `runPhasePlanners`, so `resolvePhaseEconomyDevelopActive` mirrors
-  // `resolvePhaseEconomyColonialPressureActive` (this file) and
-  // `resolvePhaseDiplomacyDeclareWarDevelopSuppressionActive`
-  // (`phase_planner_diplomacy_filter.dart`) by routing the DEVELOP
-  // gate off the dispatched value. Phase-derived `true/false` is
-  // field-equal to the legacy compute across every
-  // [ObserverGoalPhase] value, preserving the prior workThreshold cap
-  // / `runFullAiCivilianWork` behaviour exactly under DEVELOP.
   final developPhase = economyPhaseGates.developActive;
-  // Refs #2509 S5: derive colonial economy pressure from the dispatched
-  // phase plan instead of the legacy three-predicate compute. The phase
-  // dispatcher already resolved `observerGoalPhaseFor` once per player turn;
-  // this resolver mirrors `resolvePhaseConquestColonialPressureActive` and
-  // enables the colonial economy boost only under COLONIAL — structurally
-  // suppressed under EXPAND, COLONIAL-lite, and DEVELOP per
-  // `SPEC/ai/phase-planner-dispatch.md` § Orchestrator economy slice.
-  // The tagalong `newWorldProvincesOwned > 0` guards below still
-  // independently trigger the colonial workThreshold cap and
-  // `runFullAiCivilianWork` so GPs that already own NW provinces keep
-  // running civilian planning under EXPAND / COLONIAL-lite.
   final colonialPressure = economyPhaseGates.colonialPressureActive;
-  // Refs #2847 Phase 3 economy civilian-work threshold cap wiring: the
-  // colonial-pressure civilian-work cap now scales continuously with the
-  // soft-phase NW acquisition weight instead of stepping to the hard
-  // `kColonialCivilianWorkThresholdCap` only under the COLONIAL boolean.
-  // `economyColonialPressureCivilianWorkThresholdCap` is identity-equal to
-  // the legacy uncapped threshold at weight 0.0 and to the hard colonial
-  // cap at weight 1.0 (see `SPEC/ai/phase-planner-architecture.md`
-  // § Phase 3 consumer wiring — economy civilian-work threshold cap). The
-  // `newWorldProvincesOwned > 0` tagalong remains a hard cap so a GP that
-  // already holds NW land keeps the full colonial civilian-work bar.
   final colonialPressureWeight = economyPhaseGates.colonialPressureWeight;
   if (developPhase) {
     workThreshold = math.min(workThreshold, kDevelopCivilianWorkThresholdCap);
@@ -174,8 +119,6 @@ EconomyDomainPlannersResult runEconomyDomainPlanners({
       );
     }
   }
-  // Refs #3371: growth-stage economy scoring coexists with H8 civilian-work
-  // feedstock routing until AC9 replaces it with priority-vector scoring.
   final growthStage = growthStagePlannerEnabled
       ? GrowthStage.compute(ctx.game, ctx.nationId, snapshot: snapshot)
       : null;
@@ -203,79 +146,15 @@ EconomyDomainPlannersResult runEconomyDomainPlanners({
     'workCandidatesCount=${workCandidates.length}',
   );
   if (runFullAiCivilianWork) {
-    final prioritizedWorkCandidates = growthStage != null
-        ? prioritizeWorkOrdersForGrowthStage(
-            workCandidates: workCandidates,
-            game: ctx.game,
-            playerId: ctx.nationId,
-            stage: growthStage,
-          )
-        : workCandidates;
-    // Refs #3371 AC1/AC2: route bootstrap/infrastructure Builders onto fabric
-    // (then infrastructure) feedstock tiles inside the per-unit selection. The
-    // candidate reorder above is re-sorted lexicographically per unit by the
-    // selector, so the binding signal is the feedstock resource-id preference
-    // threaded into the build-improvement scoring below.
-    final feedstockPreference = growthStage != null
-        ? growthStageFeedstockPreference(
-            game: ctx.game,
-            playerId: ctx.nationId,
-            stage: growthStage,
-            growthStagePlannerEnabled: growthStagePlannerEnabled,
-          )
-        : GrowthStageFeedstockPreference.none;
-    if (growthStage != null) {
-      final relocation = suggestGrowthStageBuilderFeedstockRelocation(
-        game: ctx.game,
-        view: ctx.view,
-        topology: ctx.topology,
-        currentOrders: ordersBuilder.build(),
-        suggestionAPI: ctx.suggestionAPI,
-        stage: growthStage,
-        feedstockPreference: feedstockPreference,
-        growthStagePlannerEnabled: growthStagePlannerEnabled,
-      );
-      if (relocation != null) {
-        _log.i(
-          'growth_stage_builder_relocate nationId=${ctx.nationId} '
-          'unitId=${relocation.unitId} '
-          'destinationTileKey=${relocation.destinationTileKey}',
-        );
-        ordersBuilder.appendMoveOrders(ctx.nationId, [relocation]);
-        final movedUnitIds = {relocation.unitId};
-        workCandidates = workCandidates
-            .where((w) => !movedUnitIds.contains(w.unitId))
-            .toList();
-      }
-    }
-    final selection = selectFullAiCivilianWorkOrders(
-      workSuggestions: prioritizedWorkCandidates,
-      view: ctx.view,
-      game: ctx.game,
+    runEconomyCivilianWorkPass(
+      ctx: ctx,
+      economyPhaseGates: economyPhaseGates,
+      ordersBuilder: ordersBuilder,
+      workCandidates: workCandidates,
+      growthStage: growthStage,
+      growthStagePlannerEnabled: growthStagePlannerEnabled,
       tileMapByRegion: tileMapByRegion,
-      topology: ctx.topology,
-      growthStageFabricFeedstockResourceIds:
-          feedstockPreference.fabricFeedstockResourceIds,
-      growthStageInfraFeedstockResourceIds:
-          feedstockPreference.infraFeedstockResourceIds,
-      spyDevelopPhase: developPhase,
     );
-    for (final w in selection.workOrders) {
-      final unitType = ctx.view.ownUnitsById[w.unitId]?.type ?? 'unknown';
-      _log.i(
-        'civilian_work_assigned nationId=${ctx.nationId} unitId=${w.unitId} '
-        'unitType=$unitType target=${w.target} targetTileKey=${w.targetTileKey}',
-      );
-    }
-    for (final idle in selection.idleEvents) {
-      _log.i(
-        'civilian_work_idle nationId=${ctx.nationId} unitId=${idle.unitId} '
-        'unitType=${idle.unitType} reason=${idle.reason}',
-      );
-    }
-    if (selection.workOrders.isNotEmpty) {
-      ordersBuilder.appendWorkOrders(ctx.nationId, selection.workOrders);
-    }
   } else if (workCandidates.isNotEmpty) {
     _log.d('work skipped nationId=${ctx.nationId} weight below threshold');
   }
@@ -314,70 +193,3 @@ EconomyDomainPlannersResult runEconomyDomainPlanners({
     ),
   );
 }
-
-/// Appends a single peasant recruit-worker order into [ordersBuilder] when the
-/// growth-stage worker-growth priority (Refs #3371) or the legacy castIron
-/// labour expand boost authorizes it and the GP can afford it.
-///
-/// Extracted verbatim from [runEconomyDomainPlanners] to keep that
-/// orchestrator slice within the repo function-size budget; behaviour is
-/// unchanged.
-void appendEconomyPeasantRecruit({
-  required PlannerContext ctx,
-  required ExpandEconomyPlan expandEconomy,
-  required GrowthStage? growthStage,
-  required bool growthStagePlannerEnabled,
-  required OrdersBuilder ordersBuilder,
-}) {
-  final growthStagePeasantRecruit =
-      growthStage != null && growthStage.workerGrowthPriority > 0.1;
-  if (growthStagePeasantRecruit ||
-      (!growthStagePlannerEnabled &&
-          expandEconomy.boostCastIronLabourPeasantRecruitment)) {
-    final recruitCandidates = ctx.suggestionAPI.suggestRecruitWorkerOrders(
-      ctx.view,
-      ctx.game,
-      ctx.topology,
-      ordersBuilder.build(),
-    );
-    RecruitWorkerOrder? peasantRecruit;
-    for (final candidate in recruitCandidates) {
-      if (candidate.targetTier == WorkerTier.peasant) {
-        peasantRecruit = candidate;
-        break;
-      }
-    }
-    if (peasantRecruit != null) {
-      final player = ctx.game.playerById(ctx.nationId);
-      final affordable =
-          player != null &&
-          canAffordRecruitWorker(
-            player,
-            peasantRecruit,
-            player.workerPool,
-            player.stockpile,
-            player.treasury,
-          ).canAfford;
-      if (affordable) {
-        _log.i(
-          growthStagePeasantRecruit
-              ? 'growth-stage peasant recruit nationId=${ctx.nationId} '
-                    'workerGrowth=${growthStage.workerGrowthPriority.toStringAsFixed(2)}'
-              : 'castIron labour peasant recruit nationId=${ctx.nationId} '
-                    'targetTier=${peasantRecruit.targetTier.name}',
-        );
-        ordersBuilder.appendRecruitWorkerOrders(ctx.nationId, [peasantRecruit]);
-      } else {
-        _log.d(
-          growthStagePeasantRecruit
-              ? 'growth-stage peasant recruit deferred nationId=${ctx.nationId} '
-                    'reason=unaffordable'
-              : 'castIron labour peasant recruit deferred nationId=${ctx.nationId} '
-                    'reason=fabric_short',
-        );
-      }
-    }
-  }
-}
-
-/// Build pass outcome plus the resolved build-threshold gate decision.
